+++ packages/common/src/serializationHelpers.ts
@@ -7,7 +7,6 @@ import {
     ConfigureDelegationPayload,
     DelegationTarget,
     DelegationTargetType,
-    HexString,
 } from './types';
 import { DataBlob } from './types/DataBlob';
 import { isDefined } from './util';
@@ -294,15 +293,10 @@ function getPayloadBitmap<T>(payload: T, fieldOrder: Array<keyof T>) {
         );
 }
 
-// Makes all properties of type T non-optional.
-export type NotOptional<T> = {
-    [P in keyof T]-?: T[P];
-};
-
 /**
  * Makes a type with keys from Object and values being functions that take values with types of respective original values, returning a Buffer or undefined.
  */
-type SerializationSpec<T> = NotOptional<{
+type SerializationSpec<T> = Required<{
     [P in keyof T]: (v: T[P]) => Buffer | undefined;
 }>;
 
@@ -423,30 +417,15 @@ export function serializeConfigureBakerPayload(
 }
 
 /**
- * Helper function to create a function that deserializes a `HexString` value into a list of dynamic type values
- * determined by the deserialization logic defined in the callback function.
- *
- * @param {Function} deserializer - A callback function invoked with a `Buffer` containing the remaining slice of the full value given by the `input`
- * The callback function is expected to return the deserialized value  of type `R` along with the number of bytes processed in the format of {value: R, bytesRead: number}
- *
- * @returns {Function} A function taking a single `HexString` input, returning a list of dynamic type values deserialized according to the `deserializer` function.
+ * Takes a function which serializes a value, and returns a function that serializes an optional value
+ * Prefixed with a byte indicating if a value follows or not.
  */
-export const makeDeserializeListResponse =
-    <R>(deserializer: (buffer: Buffer) => { value: R; bytesRead: number }) =>
-    (value: HexString): R[] => {
-        const buf = Buffer.from(value, 'hex');
-        const n = buf.readUInt16LE(0);
-        let cursor = 2; // First 2 bytes hold number of token amounts included in response.
-        const values: R[] = [];
-
-        for (let i = 0; i < n; i++) {
-            const { value, bytesRead } = deserializer(
-                Buffer.from(buf.subarray(cursor))
-            );
-            values.push(value);
-
-            cursor += bytesRead;
+export const makeSerializeOptional =
+    <T>(fun: (value: T) => Buffer): ((value: T | undefined) => Buffer) =>
+    (value) => {
+        if (value === undefined) {
+            return encodeBool(false);
         }
 
-        return values;
+        return Buffer.concat([encodeBool(true), fun(value)]);
     };