+++ packages/common/src/types.ts
@@ -46,6 +46,18 @@ export type Round = bigint;
 export type Energy = bigint;
 
 /**
+ * Makes keys of type optional
+ *
+ * @example
+ * type PartiallyOptionalProps = MakeOptional<{test: string; another: number;}, 'another'>; // {test: string; another?: number;}
+ */
+export type MakeOptional<T, K extends keyof T> = Omit<T, K> &
+    Partial<Pick<T, K>>;
+
+/** Makes keys of type required (i.e. non-optional) */
+export type MakeRequired<T, K extends keyof T> = Required<Pick<T, K>> &
+    Omit<T, K>;
+/**
  * Returns a union of all keys of type T with values matching type V.
  */
 export type KeysMatching<T, V> = {