+++ packages/nodejs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## 9.2.0
+
+- Bumped @concordium/common-sdk to 9.2.0.
+
+## 9.1.1
+
+### Changed
+
+- Bumped @concordium/common-sdk to 9.1.1. (includes fixes for `verifyWeb3IdCredentialSignature` and `canProveAtomicStatement`)
+
+## 9.1.0
+
+### Changed
+
+- Bumped @concordium/common-sdk to 9.1.0. (adds methods for creating verifiable presentation (proving statements about Web3Id Credentials))
+
 ## 9.0.0
 
 ### Breaking changes